Don't post rx data buffers if not needed
If the provider does not signal FI_RX_CQ_DATA, we don't need to post
receive buffers to the endpoint. For the TCP provider this will actually
lead to receive context to overflow after 254 grain transfers, because
the buffers are not used at all. After 254 the fi_cq_read will return
-FI_EAGAIN and all subsequent reads will fail.

Signed-off-by: Jonas Ohland <jonas.ohland@riedel.net>
Co-authored-by: Jonas Ohland <jonas.ohland@gmail.com>